remove tests related to jwedb code add test for mnemonic_info
